agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
google-generativeai==0.8.5
python-dotenv==1.0.1
pytest==8.4.1
httpx==0.28.1
google-api-python-client==2.193.0
google-auth==2.49.1
